maps: Handle `GetValueAndDeleteBatch` return code appropiately
Long story short, some libbpf APIs that don't return pointers, such as the
batch APIs might return `-1` when there's a 'partial success'. In the case
of these APIs, with errno=ENOENT, we know that we read and deleted less than
`count` entries, but it was still overall sucessful.

There's more background in the inline comments). We are unfortunately
breaking the API but I think it's needed. Happy to hear your feedback! :D

This PR fixes this for `GetValueAndDeleteBatch` only. We aren't addressing
all the other batch calls yet as we haven't used them
yet.

Thanks!